allow empty queries
allows client to be used this way:

```js
index.search(function(){});
index.search(undefined, function(){});
index.search(null, function(){});
```

Because we say in our docs that it can be used this way to
get all documents (empty query) but it
did not worked previously.

Also fixed the shared cache issue. Because objects are passed as
references, creating new instances of index or client did not create a
new cache.

Prototype properties are error prone,
especially object ones where we get references.

Cache is an instance property, set it in the constructor.