chore: change the way we handle deps
Our modules can be used in two ways:
- direct npm dependent
- <script> on jsDelivr

To tackle code duplication in modules using us as a direct
dependency, I switched all direct dependent to ^x.x.x version.

This means "We are compatible with any minor/patch version you may
already have for this dependency". This allow the host project(s) to
avoid code duplication.

As for the builded files, I added an npm-shrinkwrap.json file to lock
down our dependencies.

This has a very nice advantage that now we do control the result of
the build being pushed on jsdelivr. While previously it was up to the
developer node_modules state.

Since we have an npm run release task and since it's doing `npm
install`, we will always end up with a build where the dependencies
are locked.

The dev dependencies stays the same because there's no clear win to
auto upgrade them BUT break the build :)