fix(changelog) make indentation consistent (#535)
Because the changelog wasn't being rendered as markdown before now, we didn't notice that the indentation was being inconsistent (2, 4 and 6 spaces in the same file), and thus it wouldn't render properly. This PR fixes that, and all bullet points are now using 2 spaces for indentation, which gets rendered as proper markdown.